+++ src/main/java/fr/diginamic/maison/Cuisine.java
@@ -3,7 +3,7 @@ package fr.diginamic.maison;
 // J1 Exercice Immobilier
 public class Cuisine extends Piece {
 
-    public Cuisine(double superficie, short nmrEtage) {
+    public Cuisine(double superficie, int nmrEtage) {
         super(superficie, nmrEtage);
     }
 }
