+++ src/main/java/fr/diginamic/maison/SalleDeBain.java
@@ -3,7 +3,7 @@ package fr.diginamic.maison;
 // J1 Exercice Immobilier
 public class SalleDeBain extends Piece{
 
-    public SalleDeBain(double superficie, short nmrEtage) {
+    public SalleDeBain(double superficie, int nmrEtage) {
         super(superficie, nmrEtage);
     }
 }
