+++ src/main/java/fr/diginamic/maison/Salon.java
@@ -3,7 +3,7 @@ package fr.diginamic.maison;
 // J1 Exercice Immobilier
 public class Salon extends Piece{
 
-    public Salon(double superficie, short nmrEtage) {
+    public Salon(double superficie, int nmrEtage) {
         super(superficie, nmrEtage);
     }
 }
