+++ src/main/java/fr/diginamic/maison/WC.java
@@ -3,7 +3,7 @@ package fr.diginamic.maison;
 // J1 Exercice Immobilier
 public class WC extends  Piece {
 
-    public WC(double superficie, short nmrEtage) {
+    public WC(double superficie, int nmrEtage) {
         super(superficie, nmrEtage);
     }
 }
